requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0

# server only
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

bcrypt>=4.0.0
python-dotenv